🚀(bot) Skip backup when the executable was not found

+++ bot/hooks/stop.sh
@@ -7,5 +7,9 @@ else
   echo "Directory was already created."
 fi
 
-PREFIX="$(date '+%Y%m%d-%H%M%S')-$RANDOM"
-mv /home/ec2-user/bot/gadis_provider /home/ec2/bot/executable_backups/$PREFIX-gadis_provider
+if [ -f /home/ec2-user/bot/gadis_provider ]; then
+  PREFIX="$(date '+%Y%m%d-%H%M%S')-$RANDOM"
+  mv /home/ec2-user/bot/gadis_provider /home/ec2/bot/executable_backups/$PREFIX-gadis_provider
+else
+  echo "No executables found."
+fi
